# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface GithubAccount {
  login: string;
  name: string | null;
  avatarUrl: string;
  htmlUrl: string;
}

export interface GithubRepository {
  id: number;
  name: string;
  fullName: string;
  ownerLogin: string;
  description: string | null;
  defaultBranch: string;
  private: boolean;
  htmlUrl: string;
  updatedAt: string | null;
}

export interface GithubBranch {
  name: string;
  sha: string;
  protected: boolean;
}

export interface GithubFile {
  path: string;
  content: string;
  sha: string;
}

export type GithubCommitChangeStatus = typeof GithubCommitChangeStatus[keyof typeof GithubCommitChangeStatus];


export const GithubCommitChangeStatus = {
  added: 'added',
  modified: 'modified',
  deleted: 'deleted',
} as const;

export interface GithubCommitChange {
  path: string;
  status: GithubCommitChangeStatus;
  content: string | null;
  sha: string | null;
}

export interface GithubCommitRequest {
  /** @minLength 1 */
  branch: string;
  /** @minLength 1 */
  message: string;
  /** @minItems 1 */
  changes: GithubCommitChange[];
}

export interface GithubCommitResult {
  sha: string;
  url: string;
  message: string;
}

/**
 * GitHub API error
 */
export type UpstreamErrorResponse = ApiError;

export type GetGithubFilesParams = {
/**
 * @minLength 1
 */
branch: string;
};