# lib/api-zod/src/generated/types/githubCommitChange.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubCommitChangeStatus } from './githubCommitChangeStatus';

export interface GithubCommitChange {
  path: string;
  status: GithubCommitChangeStatus;
  content: string | null;
  sha: string | null;
}