# lib/api-zod/src/generated/types/githubCommitRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubCommitChange } from './githubCommitChange';

export interface GithubCommitRequest {
  /** @minLength 1 */
  branch: string;
  /** @minLength 1 */
  message: string;
  /** @minItems 1 */
  changes: GithubCommitChange[];
}