# lib/api-zod/src/generated/types/githubFile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubFile {
  path: string;
  content: string;
  sha: string;
}